pyproject.toml
[tool.poetry]
name = "remarkable-template-backend"
version = "0.1.0"
description = ""
authors = ["Enrico Borba <enricozb@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
google-cloud-storage = "^1.31.0"
isort = "^5.4.2"
black = "^20.8b1"
fastapi = "^0.61.1"
uvicorn = "^0.11.8"
python-multipart = "^0.0.5"
Pillow = "^7.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
